Fix Crawford issues when Options/Game/Start new games immediately is off and the player analyses and browses the current game before starting the next one.
The Crawford status for the next game is set at the end of the current game to allow to display the score, for instance "4-0 to 5, Crawford game". If Start new games immediately is on, the initial record of the next game is created immediately with the right Crawford status but if this option is off looking at any move from the current game would clobber this status before the initial record of the next game could be created.

This fixes bugs #44255 and #50512 (which are duplicates, as far as I can tell) and #44037.